WIP:Work on external reference file paths
