+++ src/EPPlus/ExternalReferences/ExcelExternalDdeItemCollection.cs
@@ -10,13 +10,11 @@
  *************************************************************************************************
   04/16/2021         EPPlus Software AB       EPPlus 5.7
  *************************************************************************************************/
-using System.Collections.Generic;
+using OfficeOpenXml.Core;
 
-namespace OfficeOpenXml.Core.ExternalReferences
+namespace OfficeOpenXml.ExternalReferences
 {
-    public class ExcelExternalDdeValue
+    public class ExcelExternalDdeItemCollection : EPPlusReadOnlyList<ExcelExternalDdeItem>
     {
-        public eDdeValueType DdeValueType { get; set; } = eDdeValueType.Number;
-        public string Value { get; set; }
     }
 }
