+++ src/EPPlus/ExternalReferences/ExcelExternalDdeLink.cs
@@ -10,45 +10,47 @@
  *************************************************************************************************
   04/16/2021         EPPlus Software AB       EPPlus 5.7
  *************************************************************************************************/
+using OfficeOpenXml.Core.CellStore;
 using OfficeOpenXml.Packaging;
 using OfficeOpenXml.Utils;
 using OfficeOpenXml.Utils.Extensions;
+using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Linq;
+using System.Text;
 using System.Xml;
 
-namespace OfficeOpenXml.Core.ExternalReferences
+namespace OfficeOpenXml.ExternalReferences
 {
-    public class ExcelExternalOleLink : ExcelExternalLink
+    public class ExcelExternalDdeLink : ExcelExternalLink
     {
-        internal ExcelExternalOleLink(ExcelWorkbook wb, XmlTextReader reader, ZipPackagePart part, XmlElement workbookElement) : base(wb, reader, part, workbookElement)
+        internal ExcelExternalDdeLink(ExcelWorkbook wb, XmlTextReader reader, ZipPackagePart part, XmlElement workbookElement) : base (wb, reader, part, workbookElement)
         {
-            var rId = reader.GetAttribute("id", ExcelPackage.schemaRelationships);
-            if(!string.IsNullOrEmpty(rId))
-            {
-                Relation = part.GetRelationship(rId);
-            }
-            ProgId = reader.GetAttribute("progId");
+            DdeService = reader.GetAttribute("ddeService");
+            DdeTopic = reader.GetAttribute("ddeTopic");
+            
             while (reader.Read())
             {
                 if (reader.NodeType == XmlNodeType.Element)
                 {
                     switch (reader.LocalName)
                     {
-                        case "oleItems":
-                            ReadOleItems(reader);
+                        case "ddeItems":
+                            ReadDdeItems(reader);
                             break;
                     }
                 }
                 else if (reader.NodeType == XmlNodeType.EndElement)
                 {
-                    if (reader.Name == "oleLink")
+                    if (reader.Name == "ddeLink")
                     {
                         break;
                     }
                 }
             }
         }
-        private void ReadOleItems(XmlTextReader reader)
+        private void ReadDdeItems(XmlTextReader reader)
         {
             while (reader.Read())
             {
@@ -56,15 +58,19 @@ namespace OfficeOpenXml.Core.ExternalReferences
                 {
                     XmlStreamHelper.ReadUntil(reader, "Fallback");
                 }
-                if (reader.NodeType == XmlNodeType.Element && reader.LocalName == "oleItem")
+                if (reader.NodeType == XmlNodeType.Element && reader.LocalName == "ddeItem")
                 {
-                    OleItems.Add(new ExcelExternalOleItem()
+                    DdeItems.Add(new ExcelExternalDdeItem()
                     {
                         Name = reader.GetAttribute("name"),
                         Advise = XmlHelper.GetBoolFromString(reader.GetAttribute("advise")),
-                        Icon = XmlHelper.GetBoolFromString(reader.GetAttribute("icon")),
+                        Ole = XmlHelper.GetBoolFromString(reader.GetAttribute("ole")),
                         PreferPicture = XmlHelper.GetBoolFromString(reader.GetAttribute("preferPic")),
                     });
+                }
+                else if (reader.NodeType == XmlNodeType.EndElement && reader.LocalName=="ddeItems")
+                {
+                    break;
                 }
             }
         }
@@ -73,32 +79,29 @@ namespace OfficeOpenXml.Core.ExternalReferences
         {
             get
             {
-                return eExternalLinkType.OleLink;
+                return eExternalLinkType.DdeLink;
             }
         }
-        internal ZipPackageRelationship Relation
-        {
-            get;
-            set;
-        }
-        public ExcelExternalOleItemsCollection OleItems
-        {
-            get;
-        } = new ExcelExternalOleItemsCollection();
-        public string ProgId { get; }
+        public string DdeService { get; set; }
+        public string DdeTopic { get; set; }
 
+        public ExcelExternalDdeItemCollection DdeItems
+        {
+            get;
+        } = new ExcelExternalDdeItemCollection();
         internal override void Save(StreamWriter sw)
         {
-            sw.Write($"<oleLink progId=\"{ProgId}\" r:Id=\"{Relation.Id}\" xmlns:r=\"http://schemas.openxmlformats.org/officeDocument/2006/relationships\"><oleItems>");
-            foreach (ExcelExternalOleItem item in OleItems)
-            {
-                sw.Write(string.Format("<mc:AlternateContent><mc:Choice Requires=\"x14\"><x14:oleItem name=\"{0}\" {1}{2}{3}/></mc:Choice><mc:Fallback><oleItem name=\"{0}\" {1}{2}{3}/></mc:Fallback></mc:AlternateContent>",
+            sw.Write($"<ddeLink ddeTopic=\"{DdeTopic}\" ddeService=\"{DdeService}\"><ddeItems>");
+            foreach (ExcelExternalDdeItem item in DdeItems)
+            {                
+                sw.Write(string.Format("<ddeItem name=\"{0}\" {1}{2}{3}/>",
                   item.Name,
                   item.Advise.GetXmlAttributeValue("advise", false),
-                  item.Icon.GetXmlAttributeValue("icon", false),
+                  item.Ole.GetXmlAttributeValue("ole", false),
                   item.PreferPicture.GetXmlAttributeValue("preferPic", false)));
             }
-            sw.Write("</oleItems></oleLink>");
+            sw.Write("</ddeItems></ddeLink>");
         }
     }
 }
+
