+++ src/EPPlus/ExternalReferences/ExcelExternalDdeValue.cs
@@ -10,11 +10,13 @@
  *************************************************************************************************
   04/16/2021         EPPlus Software AB       EPPlus 5.7
  *************************************************************************************************/
+using System.Collections.Generic;
 
-namespace OfficeOpenXml.Core.ExternalReferences
+namespace OfficeOpenXml.ExternalReferences
 {
-    public interface IExcelExternalNamedItem
+    public class ExcelExternalDdeValue
     {
-        string Name { get; }
+        public eDdeValueType DdeValueType { get; set; } = eDdeValueType.Number;
+        public string Value { get; set; }
     }
 }
