+++ src/EPPlus/ExternalReferences/ExcelExternalOleItem.cs
@@ -10,30 +10,13 @@
  *************************************************************************************************
   04/16/2021         EPPlus Software AB       EPPlus 5.7
  *************************************************************************************************/
-using OfficeOpenXml.Core.CellStore;
-
-namespace OfficeOpenXml.Core.ExternalReferences
+namespace OfficeOpenXml.ExternalReferences
 {
-    public class ExcelExternalWorksheet : IExcelExternalNamedItem
+    public class ExcelExternalOleItem 
     {
-        internal ExcelExternalWorksheet(
-            CellStore<object> values,
-            CellStore<int> metaData,
-            ExcelExternalNamedItemCollection<ExcelExternalDefinedName> definedNames)
-        {
-            Names = definedNames;
-            CellValues = new ExcelExternalReferenceCellCollection(values, metaData);
-        }
-        public int SheetId { get; set; }
-        public string Name { get; internal set; }
-        public EPPlusReadOnlyList<ExcelExternalDefinedName> Names { get; }
-        public ExcelExternalReferenceCellCollection CellValues 
-        { 
-            get; 
-        }
-        public override string ToString()
-        {
-            return Name;
-        }
+        public bool Advise { get; set; }
+        public bool PreferPicture { get; set; }
+        public bool Icon { get; set; }
+        public string Name { get; set; }
     }
 }
