+++ src/EPPlus/ExternalReferences/enum/eDdeValueType.cs
@@ -10,9 +10,32 @@
  *************************************************************************************************
   04/16/2021         EPPlus Software AB       EPPlus 5.7
  *************************************************************************************************/
-namespace OfficeOpenXml.Core.ExternalReferences
+namespace OfficeOpenXml.ExternalReferences
 {
-    public class ExcelExternalOleItemsCollection : EPPlusReadOnlyList<ExcelExternalOleItem>
+    /// <summary>
+    /// The type of DDE value.
+    /// </summary>
+    public enum eDdeValueType
     {
+        /// <summary>
+        /// The value is a boolean.
+        /// </summary>
+        Boolean,
+        /// <summary>
+        /// The value is an error.
+        /// </summary>
+        Error,
+        /// <summary>
+        /// The value is a real number.
+        /// </summary>
+        Number,
+        /// <summary>
+        /// The value is nil.
+        /// </summary>
+        Nil,
+        /// <summary>
+        /// The value is a string.
+        /// </summary>
+        String
     }
 }
